+++ Mods/Automatic_Inventory_Manager/ScriptExtender/Lua/AIM/Listeners/_Initializers.lua
@@ -33,7 +33,6 @@ end
 
 Ext.Events.ResetCompleted:Subscribe(function(_)
 	if Config.AIM.ENABLED == 1 then
-		Config.SyncConfigsAndFilters()
 		ResetItemStacks()
 	end
 end)